+++ darwin/Classes/WebviewCookieManagerPlugin.h
@@ -1,4 +1,8 @@
+#if TARGET_OS_OSX
+#import <FlutterMacOS/FlutterMacOS.h>
+#else
 #import <Flutter/Flutter.h>
+#endif
 
 @interface WebviewCookieManagerPlugin : NSObject<FlutterPlugin>
 @end
