+++ darwin/Classes/WebviewCookieManagerPlugin.m
